TransactioEdits

+++ src/main/java/com/example/personal_budget/exception/TransactionNotFoundException.java
@@ -1,9 +1,8 @@
 package com.example.personal_budget.exception;
 
-import java.math.BigInteger;
 
 public class TransactionNotFoundException extends RuntimeException {
-  public TransactionNotFoundException(BigInteger id) {
+  public TransactionNotFoundException(Long id) {
     super("Transaction with id " + id + " not found");
   }
 }
